chore(tests): Split model tests into separate files
