+++ netbox_acls/tests/models/test_accesslists.py
@@ -1,79 +1,14 @@
 from itertools import cycle
 
-from dcim.models import (
-    Device,
-    DeviceRole,
-    DeviceType,
-    Interface,
-    Manufacturer,
-    Site,
-    VirtualChassis,
-)
+from dcim.models import Device
 from django.contrib.contenttypes.models import ContentType
 from django.core.exceptions import ValidationError
-from django.test import TestCase
 from ipam.models import Prefix
-from netaddr import IPNetwork
-from virtualization.models import Cluster, ClusterType, VirtualMachine, VMInterface
+from virtualization.models import VirtualMachine
 
-from netbox_acls.models import *
+from netbox_acls.models import AccessList
 
-
-class BaseTestCase(TestCase):
-    """
-    Base test case for netbox_acls models.
-    """
-
-    @classmethod
-    def setUpTestData(cls):
-        """
-        Create base data to test using including:
-          - 1 of each of the following: test site, manufacturer, device type, device role, cluster type, cluster, virtual_chassis, & virtual machine
-          - 2 devices, prefixes, 2 interfaces, and 2 vminterfaces
-        """
-
-        site = Site.objects.create(name="Site 1", slug="site-1")
-        manufacturer = Manufacturer.objects.create(
-            name="Manufacturer 1",
-            slug="manufacturer-1",
-        )
-        devicetype = DeviceType.objects.create(
-            manufacturer=manufacturer,
-            model="Device Type 1",
-        )
-        devicerole = DeviceRole.objects.create(
-            name="Device Role 1",
-            slug="device-role-1",
-        )
-        device = Device.objects.create(
-            name="Device 1",
-            site=site,
-            device_type=devicetype,
-            device_role=devicerole,
-        )
-        # virtual_chassis = VirtualChassis.objects.create(name="Virtual Chassis 1")
-        # virtual_chassis_member = Device.objects.create(
-        #    name="VC Device",
-        #    site=site,
-        #    device_type=devicetype,
-        #    device_role=devicerole,
-        #    virtual_chassis=virtual_chassis,
-        #    vc_position=1,
-        # )
-        # cluster_member = Device.objects.create(
-        #    name="Cluster Device",
-        #    site=site,
-        #    device_type=devicetype,
-        #    device_role=devicerole,
-        # )
-        # clustertype = ClusterType.objects.create(name="Cluster Type 1")
-        # cluster = Cluster.objects.create(
-        #    name="Cluster 1",
-        #    type=clustertype,
-        # )
-        virtual_machine = VirtualMachine.objects.create(name="VirtualMachine 1")
-        virtual_machine.save()
-        prefix = Prefix.objects.create(prefix="10.0.0.0/8")
+from .base import BaseTestCase
 
 
 class TestAccessList(BaseTestCase):
@@ -115,7 +50,7 @@ class TestAccessList(BaseTestCase):
 
     def test_duplicate_name_success(self):
         """
-        Test that AccessList names can be non-unique if associated to different devices.
+        Test that AccessList names can be non-unique if associated with different devices.
         """
         AccessList.objects.create(
             name="GOOD-DUPLICATE-ACL",
@@ -140,7 +75,7 @@ class TestAccessList(BaseTestCase):
 
     def test_alphanumeric_plus_fail(self):
         """
-        Test that AccessList names with non-alphanumeric (exluding '_' and '-') characters fail validation.
+        Test that AccessList names with non-alphanumeric (excluding '_' and '-') characters fail validation.
         """
         non_alphanumeric_plus_chars = " !@#$%^&*()[]{};:,./<>?\|~=+"
 
@@ -224,138 +159,3 @@ class TestAccessList(BaseTestCase):
         )
         with self.assertRaises(ValidationError):
             invalid_acl_type.full_clean()
-
-
-class TestACLInterfaceAssignment(BaseTestCase):
-    """
-    Test ACLInterfaceAssignment model.
-    """
-
-    @classmethod
-    def setUpTestData(cls):
-        """
-        Extend BaseTestCase's setUpTestData() to create additional data for testing.
-        """
-        super().setUpTestData()
-        device = Device.objects.first()
-        interfaces = Interface.objects.bulk_create(
-            (
-                Interface(name="Interface 1", device=device, type="1000baset"),
-                Interface(name="Interface 2", device=device, type="1000baset"),
-            )
-        )
-        virtual_machine = VirtualMachine.objects.first()
-        vminterfaces = VMInterface.objects.bulk_create(
-            (
-                VMInterface(name="Interface 1", virtual_machine=virtual_machine),
-                VMInterface(name="Interface 2", virtual_machine=virtual_machine),
-            )
-        )
-        # prefixes = Prefix.objects.bulk_create(
-        #    (
-        #        Prefix(prefix=IPNetwork("10.0.0.0/24")),
-        #        Prefix(prefix=IPNetwork("192.168.1.0/24")),
-        #    )
-        # )
-
-    def test_acl_interface_assignment_success(self):
-        """
-        Test that ACLInterfaceAssignment passes validation if the ACL is assigned to the host and not already assigned to the interface and direction.
-        """
-        device_acl = AccessList(
-            name="STANDARD_ACL",
-            comments="STANDARD_ACL",
-            type="standard",
-            default_action="permit",
-            assigned_object=Device.objects.first(),
-        )
-        device_acl.save()
-        acl_device_interface = ACLInterfaceAssignment(
-            access_list=device_acl,
-            direction="ingress",
-            assigned_object=Interface.objects.first(),
-        )
-        acl_device_interface.full_clean()
-
-    def test_aclinterface_assignment_fail(self):
-        """
-        Test that ACLInterfaceAssignment passes validation if the ACL is assigned to the host and not already assigned to the vminterface and direction.
-        """
-        device_acl = AccessList(
-            name="STANDARD_ACL",
-            comments="STANDARD_ACL",
-            type="standard",
-            default_action="permit",
-            assigned_object=Device.objects.first(),
-        )
-        device_acl.save()
-        acl_vm_interface = ACLInterfaceAssignment(
-            access_list=device_acl,
-            direction="ingress",
-            assigned_object=VMInterface.objects.first(),
-        )
-        with self.assertRaises(ValidationError):
-            acl_vm_interface.full_clean()
-
-    def test_acl_vminterface_assignment_success(self):
-        """
-        Test that ACLInterfaceAssignment passes validation if the ACL is assigned to the host and not already assigned to the vminterface and direction.
-        """
-        vm_acl = AccessList(
-            name="STANDARD_ACL",
-            comments="STANDARD_ACL",
-            type="standard",
-            default_action="permit",
-            assigned_object_id=1,
-            assigned_object_type=ContentType.objects.get_for_model(VirtualMachine),
-        )
-        vm_acl.save()
-        acl_vm_interface = ACLInterfaceAssignment(
-            access_list=vm_acl,
-            direction="ingress",
-            assigned_object_id=1,
-            assigned_object_type=ContentType.objects.get_for_model(VMInterface),
-        )
-        acl_vm_interface.full_clean()
-
-    def test_acl_interface_assignment_fail(self):
-        """
-        Test that ACLInterfaceAssignment fails validation if the ACL is not assigned to the parent host.
-        """
-        pass
-        #  TODO: test_acl_interface_assignment_fail - VM & Device
-
-    def test_duplicate_assignment_fail(self):
-        """
-        Test that ACLInterfaceAssignment fails validation if the ACL already is assigned to the same interface and direction.
-        """
-        pass
-        #  TODO: test_duplicate_assignment_fail - VM & Device
-
-    def test_acl_already_assinged_fail(self):
-        """
-        Test that ACLInterfaceAssignment fails validation if the interface already has an ACL assigned in the same direction.
-        """
-        pass
-        ##  TODO: test_acl_already_assinged_fail - VM & Device
-
-    # TODO: Test choices for ACLInterfaceAssignment Model
-
-
-# TODO: Investigate a Base model for ACLStandardRule and ACLExtendedRule
-
-
-class TestACLStandardRule(BaseTestCase):
-    """
-    Test ACLStandardRule model.
-    """
-
-    # TODO: Develop tests for ACLStandardRule model
-
-
-class ACLExtendedRule(BaseTestCase):
-    """
-    Test ACLExtendedRule model.
-    """
-
-    # TODO: Develop tests for ACLExtendedRule model
