+++ netbox_acls/tests/models/test_extendedrules.py
@@ -1,0 +1,9 @@
+from .base import BaseTestCase
+
+
+class ACLExtendedRule(BaseTestCase):
+    """
+    Test ACLExtendedRule model.
+    """
+
+    # TODO: Develop tests for ACLExtendedRule model
