+++ netbox_acls/tests/models/test_standardrules.py
@@ -1,0 +1,9 @@
+from .base import BaseTestCase
+
+
+class TestACLStandardRule(BaseTestCase):
+    """
+    Test ACLStandardRule model.
+    """
+
+    # TODO: Develop tests for ACLStandardRule model
